feat: abstract the TaskcatArtifactManager
Abstract the TaskcatArtifactManager class, and re-write the
post-entrypoint.ts script to instantiate dependencies and call the
class to output the taskcat log files as an artifact.
